Move the JSON variable replacer to Calamari.Common.
